move user-event out of resources
