+++ package.json
@@ -10,7 +10,7 @@
     "build": "run-s build:*",
     "build:remix": "remix build",
     "build:server": "tsx ./other/build-server.ts",
-    "dev": "remix dev -c \"npm run dev:server\"",
+    "dev": "remix dev -c \"npm run dev:server\" --no-restart",
     "dev:server": "cross-env MOCKS=true tsx watch --clear-screen=false --ignore \"app/**\" --ignore \"build/**\" --ignore \"node_modules/**\" ./index.cjs",
     "format": "prettier --write .",
     "lint": "eslint --cache --cache-location ./node_modules/.cache/eslint .",
